refactor: move concrete from vm to /concrete
