+++ cmd/geth/main.go
@@ -30,11 +30,11 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/ethereum/go-ethereum/cmd/utils"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/concrete/api"
+	cc_api "github.com/ethereum/go-ethereum/concrete/api"
+	"github.com/ethereum/go-ethereum/concrete/contracts"
+	"github.com/ethereum/go-ethereum/concrete/wasm"
 	"github.com/ethereum/go-ethereum/console/prompt"
-	"github.com/ethereum/go-ethereum/core/vm/concrete"
-	"github.com/ethereum/go-ethereum/core/vm/concrete/api"
-	cc_api "github.com/ethereum/go-ethereum/core/vm/concrete/api"
-	"github.com/ethereum/go-ethereum/core/vm/concrete/wasm"
 	"github.com/ethereum/go-ethereum/eth"
 	"github.com/ethereum/go-ethereum/eth/downloader"
 	"github.com/ethereum/go-ethereum/ethclient"
@@ -503,7 +503,7 @@ func (a *concreteGeth) Run() error {
 }
 
 func (a *concreteGeth) validateNewPCAddress(addr common.Address) error {
-	if _, ok := concrete.GetPrecompile(addr); ok {
+	if _, ok := contracts.GetPrecompile(addr); ok {
 		return fmt.Errorf("precompile already exists at address %x", addr)
 	}
 	if addr.Big().Cmp(big.NewInt(128)) < 0 {
@@ -516,14 +516,14 @@ func (a *concreteGeth) AddPrecompile(addr common.Address, pc cc_api.Precompile) 
 	if err := a.validateNewPCAddress(addr); err != nil {
 		return err
 	}
-	return concrete.AddPrecompile(addr, pc)
+	return contracts.AddPrecompile(addr, pc)
 }
 
 func (a *concreteGeth) AddPrecompileWASM(addr common.Address, code []byte) error {
 	if err := a.validateNewPCAddress(addr); err != nil {
 		return err
 	}
-	return concrete.AddPrecompile(addr, wasm.NewWasmPrecompile(code))
+	return contracts.AddPrecompile(addr, wasm.NewWasmPrecompile(code))
 }
 
 var _ ConcreteApp = &concreteGeth{}
