+++ concrete/lib/common.go
@@ -20,7 +20,7 @@ import (
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/vm/concrete/api"
+	"github.com/ethereum/go-ethereum/concrete/api"
 )
 
 func GetData(data []byte, start uint64, size uint64) []byte {
