+++ concrete/wasm/bridge/common.go
